+++ decompiler/IR2/IR2_common.h
@@ -175,7 +175,10 @@ struct UseDefInfo {
     for (auto& x : defs) {
       if (x.op_id == op_id) {
         if (x.disabled) {
-          throw std::runtime_error("Invalid disable def twice");
+          lg::warn(
+              "disable def twice: {}. This may happen when a cond (no else) is nested inside of "
+              "another conditional, but it should be rare.\n",
+              x.op_id);
         }
         x.disabled = true;
         return;