+++ doc/changelog.md
@@ -142,3 +142,4 @@
 - Adding a duplicate entry to an enum now generates a compiler error.
 - Added `.psubw` assembly form
 - Changed `.ftoi` to `VCVTTPS2DQ` to make the rounding behavior match the PS2 (truncate).
+- Forward declaring a type with `declare-type` also forward declares the global holding the runtime type object.